Move resource_usage general purpose functions to util
